+++ project/pom.xml
@@ -56,7 +56,7 @@
 		    <version>11.2.0.1.0</version>
 		</dependency>
 
-		<!-- 이메일 인증 -->
+		<!-- 이메일 인증 --> 
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-mail</artifactId>
