+++ project/src/main/java/com/example/project/config/WebConfig.java
@@ -17,16 +17,13 @@ public class WebConfig implements WebMvcConfigurer {
         registry.addViewController("/timetable").setViewName("timetable");
         registry.addViewController("/blank").setViewName("blank");
         
-        registry.addViewController("/mail").setViewName("mail");
-        registry.addViewController("/findPassword").setViewName("findPassword");
-        registry.addViewController("/findPasswordTest").setViewName("findPasswordTest");
-        
         registry.addViewController("/admin").setViewName("admin/admin");
         registry.addViewController("/admin/memberList").setViewName("admin/memberList");
         
         //registry.addViewController("/mypage").setViewName("mypage/mypage");
-        
-        
+         
+        registry.addViewController("/findPassword").setViewName("findPassword");
+        registry.addViewController("/findPasswordTest").setViewName("findPasswordTest");
         registry.addViewController("/mypage/memberProfile").setViewName("mypage/memberProfile");
         registry.addViewController("/mypage/memberUpdate").setViewName("mypage/memberUpdate");
         registry.addViewController("/mypage/memberDelete").setViewName("mypage/memberDelete");
