+++ project/src/main/java/com/example/project/controller/MypageController.java
@@ -17,71 +17,130 @@ import com.example.project.dto.UserDTO;
 import com.example.project.service.UserService;
 
 @Controller
-//@RequestMapping("/mypage")
+@RequestMapping("/mypage")
 public class MypageController {
 
 	@Autowired
-	 private UserService userService;
-	
-	@GetMapping("/mypage/memberProfile")
+	private UserService userService;
+
+	// ################## MemberProfile ##################
+
+	// 세션에 저장된 id값에 해당되는 유저의 정보 출력
+	@GetMapping("/memberProfile")
 	public String memberProfile(HttpSession session, Model model) {
-		System.out.println("마이페이지");
-		String id = (String)session.getAttribute("userId");
-		if(id != null) {
+
+		String id = (String) session.getAttribute("userId");
+		if (id != null) {
 			UserDTO userDTO = userService.getUserById(id);
 			model.addAttribute("user", userDTO);
-			System.out.println("후"+userDTO);
-			return "mypage/memberProfile";
+			return "memberProfile";
 		}
 		return "redirect:./login";
 	}
-	
-	// 수정 비밀번호체크
-	@GetMapping("/mypage/memberUpdateCheck")
+
+	// ################## MemberUpdate ##################
+
+	// 회원정보 수정 화면단
+	@GetMapping("/mypage/memberUpdate") // //mypage/member
+	public String toUpdatePage(HttpSession session, Model model) { // 회원정보 수정 페이지
+		System.out.println("회원정보 수정 페이지");
+		String id = (String) session.getAttribute("userId");
+		UserDTO userDTO = userService.getUserById(id);
+		model.addAttribute("user", userDTO);
+		return "mypage/memberUpdate";
+	}
+
+	// 회원정보 수정
+	@PostMapping("/mypage/memberUpdate")
+	public String modifyInfo(HttpSession session, UserDTO userDTO) { // 회원정보 수정
+		System.out.println("업데이트" + userDTO);
+		String id = (String) session.getAttribute("userId");
+		userDTO.setId(id);
+		System.out.println("업데이트" + userDTO);
+		userService.modifyInfo(userDTO);
+		System.out.println("컨트롤러");
+		return "redirect:./memberProfile";
+	}
+
+	// ################## MemberDelete ##################
+
+	// 회원 정탈퇴 페이지 화면단
+	@GetMapping("/memberDelete")
+	public String toDeletePage(HttpSession session, Model model) { // 회원정보 수정 페이지
+
+		System.out.println("회원정보 수정 페이지");
+		String id = (String) session.getAttribute("userId");
+		UserDTO userDTO = userService.getUserById(id);
+		model.addAttribute("user", userDTO);
+		return "memberDelete";
+
+	}
+
+	// 회원탈퇴
+	@PostMapping("/delete")
+	public String withdraw(HttpSession session) { // 탈퇴
+
+		System.out.println("탈퇴할 계정 세션을 가져옴");
+		String id = (String) session.getAttribute("userId");
+		if (id != null) {
+			userService.withdraw(id);
+			System.out.println("회원 탈퇴");
+		}
+		session.invalidate();
+		System.out.println("세션삭제완료_login창 리턴");
+		return "redirect:/login";
+	}
+
+	// ################## Password Check ##################
+
+	// 회원수정시 비밀번호 체크
+	@GetMapping("/memberUpdateCheck")
 	public String memberUpdateCheck(HttpSession session, Model model) {
-		System.out.println("비밀번호 체크페이지");
-		
-		String id = (String)session.getAttribute("userId");
-		if(id != null) {
-			return "mypage/memberUpdateCheck";
+
+		String id = (String) session.getAttribute("userId");
+		if (id != null) {
+			return "memberUpdateCheck";
 		}
 		return "redirect:./login";
 	}
-	
-	//탈퇴 비밀번호 체크
-	@GetMapping("/mypage/memberDeleteCheck")
+
+	// 회원 탈퇴시 비밀번호 체크
+	@GetMapping("/memberDeleteCheck")
 	public String memberDeleteCheck(HttpSession session, Model model) {
-		System.out.println("비밀번호 체크페이지");
-		
-		String id = (String)session.getAttribute("userId");
-		if(id != null) {
-			return "mypage/memberDeleteCheck";
+
+		String id = (String) session.getAttribute("userId");
+		if (id != null) {
+			return "memberDeleteCheck";
 		}
 		return "redirect:./login";
 	}
-	
-	
-	
-	
-	// 비밀번호 수정
-		@GetMapping("/mypage/memberPasswordUpdate")
-		public String memberPasswordUpdate(HttpSession session, Model model) {
-			System.out.println("비밀번호 변경 페이지");
-			
-			String id = (String)session.getAttribute("userId");
-			if(id != null) {
-				return "mypage/memberPasswordUpdate";
-			}
-			return "redirect:./login";
+
+	// 회원 수정/탈퇴시 비밀번호 중복 체크
+	@ResponseBody
+	@PostMapping("/pwdChk")
+	public String getpwd(UserDTO userDTO) {
+		System.out.println("아이디 컨트롤러에 POST 전달");
+
+		boolean b = userService.getpwd(userDTO);
+		System.out.println(b);
+		if (b) {
+			System.out.println("OK");
+			return "OK"; // 1이상이라는 말 => 비밀번호가 일치한다는 말
 		}
-		
-		
-		
-		
-		
-		
-		
-		
-		
-		
+		return "NO";
+	}
+
+	// ################## MemberPasswordUpdate ##################
+
+	// 비밀번호 변경 화면단
+	@GetMapping("/memberPasswordUpdate")
+	public String memberPasswordUpdate(HttpSession session, Model model) {
+
+		String id = (String) session.getAttribute("userId");
+		if (id != null) {
+			return "memberPasswordUpdate";
+		}
+		return "redirect:./login";
+	}
+
 }
