--- project/src/main/java/com/example/project/controller/UserProfileController.java
@@ -1,7 +1,0 @@
-package com.example.project.controller;
-
-public class UserProfileController {
-	
-	
-
-}
