+++ project/src/main/java/com/example/project/dto/MailDTO.java
@@ -12,3 +12,4 @@ public class MailDTO {
     private String title;
     private String message;
 }
+
