+++ project/src/main/resources/application.properties
@@ -22,4 +22,4 @@ spring.mail.port=465
 spring.mail.username=amos5105@naver.com
 spring.mail.password=Wkddbwhd@1
 spring.mail.properties.mail.smtp.auth=true
-spring.mail.properties.mail.smtp.ssl.enable=true
+spring.mail.properties.mail.smtp.ssl.enable=true
