+++ project/src/main/resources/com/config/UserMapper.xml
@@ -70,7 +70,7 @@
   		 	SET PWD = #{pwd}
   		 	WHERE ID = #{id}
   		</insert>
-  		
+  		 
   		
    
   </mapper>
